broker front end
Initial implementation for the front end of the data store.  This bit
is the interface to work with the back end data store.  Functional glue
to the rest of the app in that it provides a single function interface
that interacts with a local /table/ which updates the back end as
things change.